:recycle: jenkins_worker_fleets: remove jenkins module - it's hidding the jenkins-job-builder jenkins package
